Spot-Light/requirements.txt
PyYAML==6.0.1
openai==1.23.6
torch==2.0.1
numpy==1.24.4
wandb==0.15.8
tqdm==4.65.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.0.3
torchvision==0.15.2
bosdyn-api==4.0.2
bosdyn-choreography-client==4.0.2
bosdyn-choreography-protos==4.0.2
bosdyn-client==4.0.2
bosdyn-core==4.0.2
bosdyn-mission==4.0.2
black==24.3.0
isort==5.12.0
pylint==2.17.5
open3d==0.18.0
open3d-cpu==0.18.0
Pillow==10.0.0
scipy==1.10.0
matplotlib==3.7.2
apriltag==0.0.16
ikpy==3.3.4
Deprecated==1.2.14
pdf2image==1.16.3
Flask==2.2.5
requests==2.31.0
transformers==4.40.1
urllib3==2.0.4
scikit-learn==1.3.0
cmake==3.28.3
#git+https://github.com/openai/CLIP.git
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
matplotlib==3.7.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
scikit-image==0.21.0
scikit-learn==1.3.0
supervision==0.20.0
ultralytics==8.2.5
